Tkd-main/requirements.txt
alembic==1.13.1
appnope==0.1.3
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
comm==0.2.0
cryptolib==0.0.2
debugpy==1.8.0
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.5.1
lib==4.0.0
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
paho-mqtt==1.6.0
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyMySQL==1.1.0
python-dateutil==2.8.2
pyzmq==25.1.2
six==1.16.0
SQLAlchemy==2.0.28
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.10.0
wcwidth==0.2.12
Werkzeug==3.0.1
WTForms==3.1.2
